+++ src/main/java/com/example/transaction/application/interfaces/JwtService.java
@@ -2,5 +2,5 @@ package com.example.transaction.application.interfaces;
 
 
 public interface JwtService {
-    Long extractUserId(String token);
+    Integer extractUserId(String token);
 }
